Add option to hide/show password in login forms
This is the main logic to review in the PR.
I decided to add the eye that toggles seeing the password to the
label, rather than the input field itself. This was for two reasons.
The main reason being, I don't have a whole lot of time for this feature,
and this was the quckest way. The other reason I went with this approach is
it is very minimalistic, like PocketBase and didn't clash with the existing
design.