feat: allow the admin ui to be "detached"
This change allows users to customize if the admin UI is embeded at build time
or specified as a detached directory which exists outside the binary. This allows
users to provide a customized admin UI which is provided at runtime. Additionally
it offers the ability to reduce the binary size for users who do not care about
running the admin UI in production.

The coniditional embedding support is provided through build tags. By default the
admin UI is embeded, users can specify a detached admin UI with by building the project
with the `detachableadmin` build tag and specifying the directory for their detached
admin UI in the `PB_ADMIN_UI_DIR` env variable.