Adds Registry.ReloadString
This is a version of LoadString which busts the cache store.

A test is included that showcases an issue that you can run into, when
using AddFuncs, which is an unfortunate side-effect of Go's own template
package requiring Funcs to explicitly be called before Parse, resulting
in caching template functions as well in the current implementation.