Generate webp thumbnails
This adds support for webp thumbnails by explicitly importing the
official (but experimental) x/image/webp decoder for webp.

Unless the decoder panics (can it?), this is safe. Pocketbase will fall
back to reusing the fullsize image if creating a thumbnail fails. Which
is already the current behavior for all webp images.

Since disintegration/imaging is not aware of webp, webp thumbnails will
be generated in png format. This is a caveat but modern web browsers
generally don’t rely on file extensions to detect image formats.